Add risk acknowledgment components into views
