+++ packages/app/src/domain/common/risk.ts
@@ -1,10 +1,9 @@
 import BigNumber from 'bignumber.js'
-import { NormalizedUnitNumber } from '../types/NumericValues'
-import { Token } from '../types/Token'
 
 export type RiskLevel = 'liquidation' | 'risky' | 'moderate' | 'healthy' | 'unknown' | 'no debt'
 
 export const LIQUIDATION_HEALTH_FACTOR_THRESHOLD = new BigNumber(1)
+export const LIQUIDATION_DANGER_HEALTH_FACTOR_THRESHOLD = new BigNumber(1.2)
 export const RISKY_HEALTH_FACTOR_THRESHOLD = new BigNumber(2)
 export const MODERATE_HEALTH_FACTOR_THRESHOLD = new BigNumber(3)
 
@@ -38,33 +37,4 @@ export const riskLevelToTitle: Record<RiskLevel, string> = {
   moderate: 'Moderate',
   healthy: 'Healthy',
   'no debt': 'No debt',
-}
-
-export type RiskWarning =
-  | {
-      type: 'savings-deposit-discrepancy-threshold-hit'
-      discrepancy: NormalizedUnitNumber
-      token: Token
-    }
-  | {
-      type: 'savings-withdraw-discrepancy-threshold-hit'
-      discrepancy: NormalizedUnitNumber
-      token: Token
-    }
-  | {
-      type: 'liquidation-warning-borrow'
-    }
-  | {
-      type: 'liquidation-warning-withdraw'
-    }
-  | {
-      type: 'liquidation-warning-set-collateral'
-    }
-  | {
-      type: 'liquidation-warning-e-mode-off'
-    }
-
-export interface RiskAcknowledgementInfo {
-  onStatusChange: (acknowledged: boolean) => void
-  warning?: RiskWarning
 }
