+++ packages/app/src/features/dialogs/collateral/views/CollateralView.tsx
@@ -1,13 +1,13 @@
-import BigNumber from 'bignumber.js'
-
 import { TokenWithBalance } from '@/domain/common/types'
 import { SetUseAsCollateralValidationIssue } from '@/domain/market-validators/validateSetUseAsCollateral'
+import { RiskAcknowledgementInfo } from '@/domain/risk-warning/types'
 import { Objective } from '@/features/actions/logic/types'
 import { DialogActionsPanel } from '@/features/dialogs/common/components/DialogActionsPanel'
 import { MultiPanelDialog } from '@/features/dialogs/common/components/MultiPanelDialog'
 import { PageStatus } from '@/features/dialogs/common/types'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
-
+import { RiskAcknowledgement } from '@/ui/organisms/risk-acknowledgement/RiskAcknowledgement'
+import BigNumber from 'bignumber.js'
 import { CollateralAlert } from '../components/CollateralAlert'
 import { CollateralOverviewPanel } from '../components/CollateralOverviewPanel'
 import { CollateralSetting } from '../types'
@@ -20,6 +20,7 @@ interface CollateralViewProps {
   pageStatus: PageStatus
   currentHealthFactor?: BigNumber
   updatedHealthFactor?: BigNumber
+  riskAcknowledgement: RiskAcknowledgementInfo
 }
 
 export function CollateralView({
@@ -30,6 +31,7 @@ export function CollateralView({
   pageStatus,
   currentHealthFactor,
   updatedHealthFactor,
+  riskAcknowledgement,
 }: CollateralViewProps) {
   return (
     <MultiPanelDialog>
@@ -42,6 +44,13 @@ export function CollateralView({
         currentHealthFactor={currentHealthFactor}
         updatedHealthFactor={updatedHealthFactor}
       />
+
+      {riskAcknowledgement.warning && (
+        <RiskAcknowledgement
+          onStatusChange={riskAcknowledgement.onStatusChange}
+          warning={riskAcknowledgement.warning}
+        />
+      )}
 
       <DialogActionsPanel
         objectives={objectives}
