+++ packages/app/src/features/dialogs/savings/deposit/logic/useSavingsDepositDialog.ts
@@ -1,7 +1,7 @@
 import { getChainConfigEntry } from '@/config/chain'
-import { RiskAcknowledgementInfo } from '@/domain/common/risk'
 import { TokenWithBalance, TokenWithValue } from '@/domain/common/types'
 import { useMarketInfo } from '@/domain/market-info/useMarketInfo'
+import { RiskAcknowledgementInfo } from '@/domain/risk-warning/types'
 import { useSavingsInfo } from '@/domain/savings-info/useSavingsInfo'
 import { makeAssetsInWalletList } from '@/domain/savings/makeAssetsInWalletList'
 import { Token } from '@/domain/types/Token'
