+++ packages/app/src/features/dialogs/savings/withdraw/logic/generateWarning.ts
@@ -1,6 +1,6 @@
-import { RiskWarning } from '@/domain/common/risk'
 import { SwapInfo } from '@/domain/exchanges/types'
 import { MarketInfo } from '@/domain/market-info/marketInfo'
+import { RiskWarning } from '@/domain/risk-warning/types'
 import { SavingsInfo } from '@/domain/savings-info/types'
 import { NormalizedUnitNumber } from '@/domain/types/NumericValues'
 import { DialogFormNormalizedData } from '@/features/dialogs/common/logic/form'
