+++ packages/app/src/features/dialogs/savings/withdraw/logic/useSavingsWithdrawDialog.ts
@@ -1,8 +1,8 @@
 import { getChainConfigEntry } from '@/config/chain'
-import { RiskAcknowledgementInfo } from '@/domain/common/risk'
 import { TokenWithBalance, TokenWithValue } from '@/domain/common/types'
 import { useConditionalFreeze } from '@/domain/hooks/useConditionalFreeze'
 import { useMarketInfo } from '@/domain/market-info/useMarketInfo'
+import { RiskAcknowledgementInfo } from '@/domain/risk-warning/types'
 import { useSavingsInfo } from '@/domain/savings-info/useSavingsInfo'
 import { makeAssetsInWalletList } from '@/domain/savings/makeAssetsInWalletList'
 import { useWalletInfo } from '@/domain/wallet/useWalletInfo'
