+++ packages/app/src/features/dialogs/withdraw/views/WithdrawView.tsx
@@ -1,6 +1,5 @@
-import { UseFormReturn } from 'react-hook-form'
-
 import { TokenWithBalance, TokenWithValue } from '@/domain/common/types'
+import { RiskAcknowledgementInfo } from '@/domain/risk-warning/types'
 import { Objective } from '@/features/actions/logic/types'
 import { DialogActionsPanel } from '@/features/dialogs/common/components/DialogActionsPanel'
 import { FormAndOverviewWrapper } from '@/features/dialogs/common/components/FormAndOverviewWrapper'
@@ -9,7 +8,8 @@ import { DialogForm } from '@/features/dialogs/common/components/form/DialogForm
 import { AssetInputSchema } from '@/features/dialogs/common/logic/form'
 import { FormFieldsForDialog, PageStatus } from '@/features/dialogs/common/types'
 import { DialogTitle } from '@/ui/atoms/dialog/Dialog'
-
+import { RiskAcknowledgement } from '@/ui/organisms/risk-acknowledgement/RiskAcknowledgement'
+import { UseFormReturn } from 'react-hook-form'
 import { WithdrawOverviewPanel } from '../components/WithdrawOverviewPanel'
 import { PositionOverview } from '../logic/types'
 
@@ -22,6 +22,7 @@ export interface WithdrawViewProps {
   pageStatus: PageStatus
   currentPositionOverview: PositionOverview
   updatedPositionOverview?: PositionOverview
+  riskAcknowledgement: RiskAcknowledgementInfo
 }
 
 export function WithdrawView({
@@ -33,6 +34,7 @@ export function WithdrawView({
   withdrawAsset,
   currentPositionOverview,
   updatedPositionOverview,
+  riskAcknowledgement,
 }: WithdrawViewProps) {
   return (
     <MultiPanelDialog>
@@ -46,6 +48,13 @@ export function WithdrawView({
           updatedPositionOverview={updatedPositionOverview}
         />
       </FormAndOverviewWrapper>
+
+      {riskAcknowledgement.warning && (
+        <RiskAcknowledgement
+          onStatusChange={riskAcknowledgement.onStatusChange}
+          warning={riskAcknowledgement.warning}
+        />
+      )}
 
       <DialogActionsPanel
         objectives={objectives}
